move file to outer scope
