+++ src/app/pages/new-poem/step-1/ProsodyPage.ts
@@ -8,7 +8,7 @@ import { ReturnButton } from "../shared/ReturnButton"
 
 // Services
 import { Forms } from "../../../services/forms.service"
-import { Hint } from "../shared/Hint";
+import { Hint } from "../../shared/Hint";
 import { Title } from "../../shared/Title"
 
 interface IsStanza {
