+++ src/app/pages/shared/Hint.ts
